* Common/Histogram.h (mean, variance): Use a long unsigned rather than an unsigned to accumulate the sum and sum of squares.
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@51575 ad02020f-2a26-4133-947e-1829317c1416

+++ Common/Histogram.h
@@ -62,7 +62,7 @@ class Histogram : std::map<int, unsigned>
 
 	double mean() const
 	{
-		unsigned n = 0, total = 0;
+		long unsigned n = 0, total = 0;
 		for (Histogram::Map::const_iterator it = begin();
 				it != end(); it++) {
 			n += it->second;
@@ -73,7 +73,7 @@ class Histogram : std::map<int, unsigned>
 
 	double variance() const
 	{
-		unsigned n = 0, total = 0, squares = 0;
+		long unsigned n = 0, total = 0, squares = 0;
 		for (Histogram::Map::const_iterator it = begin();
 				it != end(); it++) {
 			n += it->second;
